requirements.txt
Pillow
Flask
pandas
python-docx
requests
multipledispatch